test(ui): fold viewport loop coverage into interactions
